# public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Transaction API",
        "contact": {
            "name": "DigitaleFactuur",
            "url": "https://digitalefactuur.nl/"
        },
        "version": "1.0"
    },
    "host": "transaction.test",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/transaction-in": {
            "get": {
                "tags": [
                    "transaction_in"
                ],
                "summary": "Array of transaction_in",
                "description": "Returns transactions array.",
                "operationId": "api.transaction_in.index",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "transaction_in[]."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "transaction_in"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns transactions overview.",
                "operationId": "api.transaction_in.store",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "account_id": {
                                    "description": "Display a listing of the resource.",
                                    "type": "integer"
                                },
                                "state_id": {
                                    "description": "Display a listing of the resource.",
                                    "type": "integer"
                                },
                                "amount": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                },
                                "date": {
                                    "description": "Display a listing of the resource.",
                                    "type": "date"
                                },
                                "origin": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/transaction-in/{id}": {
            "get": {
                "tags": [
                    "transaction_in"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns transactions object.",
                "operationId": "api.transaction_in.show",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of transaction_in to return",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "transaction_in"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns transaction overview.",
                "operationId": "api.transaction_in.delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "State id to delete",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "State overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "patch": {
                "tags": [
                    "transaction_in"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns transaction-in object that has been updated.",
                "operationId": "api.transaction_in.update",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of transaction-in to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Updated transaction object",
                        "required": true,
                        "schema": {
                            "properties": {
                                "account_id": {
                                    "description": "Display a listing of the resource.",
                                    "type": "integer"
                                },
                                "state_id": {
                                    "description": "Display a listing of the resource.",
                                    "type": "integer"
                                },
                                "amount": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                },
                                "description": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                },
                                "date": {
                                    "description": "Display a listing of the resource.",
                                    "type": "date"
                                },
                                "origin": {
                                    "description": "Display a listing of the resource.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}